moved bond flow to its own file
